backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
azure-ai-formrecognizer==3.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
google-generativeai==0.3.2
pymongo==4.6.1
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
